Smart staging algorithm for Kerbal Space Program
This mod is activated in the VAB and adds a button to automatically
compute staging calculation.
The idea is to activate as many engines as possible, follow fuel flow
and decouple as soon as all fuel is consumed.
The player just has to place the parts as needed, including decouplers
and fuel lines, press the "Smart stage" button and ideally optimal
staging would be calculated.